refactor: :recycle: refactor the draggable drawer and dependencies
